+++ vector/src/main/java/im/vector/app/core/dialogs/GalleryOrCameraDialogHelperFactory.kt
@@ -19,7 +19,11 @@ class GalleryOrCameraDialogHelperFactory @Inject constructor(
         private val colorProvider: ColorProvider,
         private val clock: Clock,
 ) {
-    fun create(fragment: Fragment): GalleryOrCameraDialogHelper {
-        return GalleryOrCameraDialogHelper(fragment, colorProvider, clock)
+    fun create(
+            fragment: Fragment,
+            aspect: GalleryOrCameraDialogHelper.Aspect = GalleryOrCameraDialogHelper.Aspect.SQUARE,
+            listener: GalleryOrCameraDialogHelper.Listener? = null,
+    ): GalleryOrCameraDialogHelper {
+        return GalleryOrCameraDialogHelper(fragment, colorProvider, clock, aspect, listener)
     }
 }